Moved the reports and main page into a separate package
Will make it a bit more clear where they are and make future
improvements easier.
