+++ src/main/resources/com/sonyericsson/hudson/plugins/gerrit/trigger/diagnostics/Diagnostics/index.properties
